backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
numpy
nltk
youtube_search
ultralytics
python-multipart
Pillow
ollama
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
bcrypt==4.0.1
